Remove RTS workaround
