--- dataprovider-retrofit/src/main/java/ch/srg/dataProvider/integrationlayer/request/image/IlHostImageUrlDecorator.kt
@@ -1,54 +1,0 @@
-package ch.srg.dataProvider.integrationlayer.request.image
-
-import android.net.Uri
-import ch.srg.dataProvider.integrationlayer.data.ImageUrlDecorator
-import ch.srg.dataProvider.integrationlayer.request.IlHost
-
-/**
- * Il host image url decorator
- *
- * If the image url isn't supported by [IlHostImageUrlDecorator] the same url is returned.
- *
- * confluence doc : https://srgssr-ch.atlassian.net/wiki/spaces/SRGPLAY/pages/799082429/Project+-+Image+Service)
- *
- * @param ilHost The [IlHost] of the integration layer image service.
- */
-class IlHostImageUrlDecorator(ilHost: IlHost) : ImageUrlDecorator {
-    private val imageServiceUri: Uri
-
-    init {
-        imageServiceUri = ilHost.hostUri.buildUpon().appendEncodedPath(IMAGES_SEGMENT).build()
-    }
-
-    override fun decorate(sourceUrl: String, widthPixels: Int): String {
-        // Il image service only support some image url hostnames!
-        if (!isImageUrlHostCompatible(sourceUrl)) return sourceUrl
-        // Il image service only support a limited image size!
-        val imageWidth = ImageWidth.getFromPixels(widthPixels)
-        return imageServiceUri.buildUpon()
-            .appendQueryParameter(PARAM_IMAGE_URL, sourceUrl)
-            .appendQueryParameter(PARAM_FORMAT, FORMAT_WEBP)
-            .appendQueryParameter(PARAM_WIDTH, imageWidth.widthPixels.toString())
-            .build()
-            .toString()
-    }
-
-    /**
-     * Check that the host of the [imageUrl] is compatible with the il image service.
-     *
-     * @param imageUrl The image url to decorate.
-     */
-    fun isImageUrlHostCompatible(imageUrl: String): Boolean {
-        return Uri.parse(imageUrl)?.host?.contains(SUPPORTED_HOST_NAME_REGEX) ?: false
-    }
-
-    companion object {
-        private const val FORMAT_WEBP = "webp" // webp, jpg, png
-        private const val IMAGES_SEGMENT = "images/"
-        private const val PARAM_IMAGE_URL = "imageUrl"
-        private const val PARAM_FORMAT = "format"
-        private const val PARAM_WIDTH = "width"
-
-        private val SUPPORTED_HOST_NAME_REGEX = "((rts|srf|rsi|rtr|swissinfo|srgssr)\\.ch)|swi-services-ch".toRegex(RegexOption.IGNORE_CASE)
-    }
-}
