--- dataprovider-retrofit/src/main/java/ch/srg/dataProvider/integrationlayer/request/image/ScaleWidthImageUrlDecorator.kt
@@ -1,23 +1,0 @@
-package ch.srg.dataProvider.integrationlayer.request.image
-
-import android.net.Uri
-import ch.srg.dataProvider.integrationlayer.data.ImageUrlDecorator
-
-/**
- * Scale width image url decorator
- *
- * @constructor Create empty Scale width image url decorator
- */
-object ScaleWidthImageUrlDecorator : ImageUrlDecorator {
-    private const val Scale = "scale"
-    private const val Width = "width"
-
-    override fun decorate(sourceUrl: String, widthPixels: Int): String {
-        return Uri.parse(sourceUrl).buildUpon()
-            .appendPath(Scale)
-            .appendPath(Width)
-            .appendPath(widthPixels.toString())
-            .build()
-            .toString()
-    }
-}
